src/core/arbitrage.ts
import { Connection, PublicKey } from '@solana/web3.js';
import { Logger } from '../utils/logger';
import { ArbitrageConfig, CustomError, PriceData, TradeOpportunity } from './types';
import { SolanaService } from './solana';
import EventEmitter from 'events';

export class ArbitrageMonitor extends EventEmitter {
  private solana: SolanaService;
  private config: ArbitrageConfig;
  private isRunning: boolean = false;

  constructor(config: ArbitrageConfig) {
    super();
    this.config = config;
    this.solana = new SolanaService(config);
  }

  async getPrices(pair: string): Promise<PriceData[]> {
    // Implement actual DEX price fetching here
    // This is a mock implementation for demonstration
    return [
      {
        dex: 'Raydium',
        pair: {
          tokenA: new PublicKey('So11111111111111111111111111111111111111112'),
          tokenB: new PublicKey('EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v')
        },
        price: 100 + Math.random() * 2,
        timestamp: Date.now()
      },
      {
        dex: 'Orca',
        pair: {
          tokenA: new PublicKey('So11111111111111111111111111111111111111112'),
          tokenB: new PublicKey('EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v')
        },
        price: 100 + Math.random() * 2,
        timestamp: Date.now()
      }
    ];
  }

  findArbitrageOpportunities(prices: PriceData[]): TradeOpportunity | null {
    if (prices.length < 2) return null;

    let bestOpportunity: TradeOpportunity | null = null;
    let maxProfit = 0;

    for (let i = 0; i < prices.length; i++) {
      for (let j = i + 1; j < prices.length; j++) {
        const buyPrice = prices[i].price;
        const sellPrice = prices[j].price;
        const profit = ((sellPrice - buyPrice) / buyPrice) * 100;

        if (profit > this.config.minProfitThreshold && profit > maxProfit) {
          maxProfit = profit;
          bestOpportunity = {
            buyDex: prices[i].dex,
            sellDex: prices[j].dex,
            profit,
            volume: 1, // Mock volume
            timestamp: Date.now()
          };
        }
      }
    }

    return bestOpportunity;
  }

  onOpportunity(callback: (opportunity: TradeOpportunity) => void) {
    this.on('opportunity', callback);
  }

  async start(pairs: string[], interval: number = 5, threshold?: number) {
    if (this.isRunning) {
      Logger.warn('Monitor is already running');
      return;
    }

    this.isRunning = true;
    if (threshold) {
      this.config.minProfitThreshold = threshold;
    }

    while (this.isRunning) {
      try {
        for (const pair of pairs) {
          const prices = await this.getPrices(pair);
          const opportunity = this.findArbitrageOpportunities(prices);

          if (opportunity) {
            this.emit('opportunity', opportunity);
          }
        }
      } catch (err) {
        Logger.error(`Error monitoring prices: ${(err as CustomError).message}`);
      }

      await new Promise(resolve => setTimeout(resolve, interval * 1000));
    }
  }

  stop() {
    this.isRunning = false;
    Logger.info('Stopping price monitor...');
  }
}